+++ addons/account_budget/__openerp__.py
@@ -33,9 +33,9 @@ Three reports are available:
     'data': [
         'security/ir.model.access.csv',
         'security/account_budget_security.xml',
-        'account_budget_view.xml',
-        'account_budget_report.xml',
-        'account_budget_workflow.xml',
+        'views/account_budget_view.xml',
+        'report/account_budget_report.xml',
+        'views/account_budget_workflow.xml',
         'wizard/account_budget_analytic_view.xml',
         'wizard/account_budget_report_view.xml',
         'wizard/account_budget_crossovered_summary_report_view.xml',
@@ -45,10 +45,10 @@ Three reports are available:
         'views/report_budget.xml',
         'views/report_crossoveredbudget.xml',
     ],
-    'demo': ['account_budget_demo.xml', 'account_budget_demo.yml'],
+    'demo': ['data/account_budget_demo.xml', 'test/account_budget_demo.yml'],
     'test': [
         '../account/test/account_minimal_test.xml',
-        'account_budget_demo.yml',
+        'test/account_budget_demo.yml',
         'test/account_budget.yml',
         'test/account_budget_report.yml',
     ],
